+++ docs/ui/mdx.tsx
@@ -3,6 +3,10 @@
 import { useMDXComponent } from 'next-contentlayer/hooks';
 import { HTMLAttributes } from 'react';
 
+import Image from 'next/image';
+
+import { cn } from '@marigold/system';
+
 import { IconList } from '@/ui/IconList';
 
 import { Headline, Link, Message, Tabs, Text } from './';
@@ -64,12 +68,16 @@ const typography = {
   // `raw` is source code to be copied
   pre: ({
     raw,
+    className,
     ...props
   }: HTMLAttributes<HTMLPreElement> & { raw: string }) => {
     return (
-      <div className="relative ">
+      <div className="relative">
         <pre
-          className="max-h-[650px] overflow-x-auto rounded-lg px-3 py-4 [&>code]:bg-transparent"
+          className={cn(
+            'max-h-[650px] overflow-x-auto rounded-lg px-3 py-4 [&>code]:bg-transparent',
+            className
+          )}
           {...props}
         >
           <div className="absolute right-4 top-4">
@@ -106,6 +114,7 @@ const components = {
   PropsTable,
   Spacing,
   TextAlign,
+  Image,
 };
 
 // Props
@@ -119,5 +128,5 @@ interface MdxProps {
 // ---------------
 export const Mdx = ({ title, code }: MdxProps) => {
   const Component = useMDXComponent(code, { title });
-  return <Component components={components} />;
+  return <Component components={components as any} />;
 };