+++ packages/components/src/HelpText/HelpText.test.tsx
@@ -1,4 +1,4 @@
-import { screen, within } from '@testing-library/react';
+import { screen } from '@testing-library/react';
 
 import { Theme, cva } from '@marigold/system';
 
@@ -63,54 +63,8 @@ test('uses description base styles', () => {
     />
   );
 
-  const element = screen.getByTestId('help-text');
-  expect(element.className).toMatchInlineSnapshot(`"flex items-center gap-1"`);
+  const element = screen.getByText('This is a help text description');
+  expect(element.className).toMatchInlineSnapshot(`"react-aria-Text"`);
 });
 
-test('renders error message when error is set', () => {
-  render(
-    <HelpText
-      error={true}
-      description="This is a help text description"
-      errorMessage="Something went wrong"
-    />
-  );
-
-  const error = screen.getByText('Something went wrong');
-  expect(error).toBeInTheDocument();
-
-  const descrption = screen.queryByText('This is a help text description');
-  expect(descrption).not.toBeInTheDocument();
-});
-
-test('renders icon when when error message is shown', () => {
-  render(
-    <HelpText
-      data-testid="help-text"
-      error={true}
-      description="This is a help text description"
-      errorMessage="Something went wrong"
-    />
-  );
-
-  const element = screen.getByTestId('help-text');
-  const icon = within(element).getByRole('presentation');
-  expect(icon).toBeInTheDocument();
-});
-
-test('icon styles via theme', () => {
-  render(
-    <HelpText
-      data-testid="help-text"
-      error={true}
-      size="small"
-      description="This is a help text description"
-      errorMessage="Something went wrong"
-    />
-  );
-
-  const element = screen.getByTestId('help-text');
-  const icon = within(element).getByRole('presentation');
-
-  expect(icon.getAttribute('class')).toMatchInlineSnapshot(`"h-3 w-3"`);
-});
+// don't tested error messages because it can't be shown (react-aria-components)