+++ packages/theme-preset/src/create-preset.ts
@@ -1,4 +1,5 @@
 import deepmerge from 'deepmerge';
+import racSelectorPlugin from 'tailwindcss-react-aria-components';
 import plugin from 'tailwindcss/plugin';
 import type { Config, OptionalConfig } from 'tailwindcss/types/config';
 
@@ -15,6 +16,8 @@ export const createPreset = (name: string, config: Partial<OptionalConfig>) =>
         },
       },
       plugins: [
+        // @ts-expect-error (are the typings wrong?)
+        racSelectorPlugin({ prefix: 'rac' }),
         // Grouping
         plugin(({ matchVariant, e }) => {
           matchVariant(