+++ themes/theme-b2b/src/components/DateField.styles.ts
@@ -1,25 +1,24 @@
 import { ThemeComponent, cva } from '@marigold/system';
 
-import {
-  inputBackground,
-  inputBox,
-  inputDisabled,
-  inputFocus,
-  inputSpacing,
-} from './Input.styles';
+import { inputBackground, inputBox, inputSpacing } from './Input.styles';
 
 export const DateField: ThemeComponent<'DateField'> = {
   field: cva([
     inputBox,
     inputBackground,
-    inputDisabled('group-disabled/field'),
-    inputFocus('group-focus/field'),
     inputSpacing,
     'h-[34px]',
-    'group-hover/field:border-border-hover',
-    'group-readonly/field:group-focus/field:border-border-light group-readonly/field:group-focus/field:outline-none',
-    ' group-readonly/field:bg-bg-transparent group-readonly/field:text-text-disabled',
+    'data-[hovered]:border-border-hover',
+    // inputDisabled('data-[disabled]') color tokens are not working with in this component
+    'data-[disabled]:bg-bg-disabled data-[disabled]:border-border-disabled data-[disabled]:text-text-disabled data-[disabled]::cursor-not-allowed',
+    // inputFocus('data-[focus-within]') is not working with this component
+    'data-[focus-within]:border-border-focus data-[focus-within]:outline-outline-focus',
   ]),
-  segment: cva(['rounded-sm', 'focus:bg-bg-focus focus:text-white']),
+  segment: cva([
+    'rounded-sm',
+    'focus:bg-bg-focus focus:text-white',
+    'aria-[readonly]:focus:border-border-light aria-[readonly]:focus:outline-none',
+    'aria-[readonly]:bg-bg-transparent aria-[readonly]:text-text-disabled',
+  ]),
   action: cva(''),
 };