+++ themes/theme-b2b/src/components/DatePicker.styles.ts
@@ -1,12 +1,12 @@
 import { ThemeComponent, cva } from '@marigold/system';
 
 export const DatePicker: ThemeComponent<'DatePicker'> = {
-  container: cva(),
-  button: cva([
-    'h-full w-9 p-0',
-    'flex items-center justify-center ',
-    'text-text-body/80',
-    'group-focus/field:[&[aria-expanded=true]]:bg-bg-secondary focus-visible:border-none focus-visible:outline-0 group-focus/field:[&[aria-expanded=true]]:text-white',
-    'disabled:hidden',
+  container: cva([
+    '[&_button]:absolute [&_button]:right-0 [&_button]:top-0 [&_button]:h-full [&_button]:w-9 [&_button]:p-0',
+    '[&_button]:flex [&_button]:items-center [&_button]:justify-center ',
+    '[&_button]:text-text-body/80',
+    '[&_button[aria-expanded=true]]:bg-bg-secondary [&_button[aria-expanded=true]]:text-white [&_button]:focus-visible:border-none [&_button]:focus-visible:outline-0',
+    '[&_button]:disabled:hidden',
   ]),
+  button: cva(),
 };