+++ themes/theme-b2b/src/theme.ts
@@ -1,9 +1,9 @@
 import { Theme } from '@marigold/system';
 
+import { colors } from './_tokens';
 import * as components from './components';
 import { root } from './root';
 import { screens } from './screens';
-import { colors } from './tokens';
 
 export const webFontUrl = [
   'https://fonts.bunny.net/css?family=inter:400,600,700',