+++ themes/theme-core/src/components/HelpText.styles.ts
@@ -1,6 +1,8 @@
 import { ThemeComponent, cva } from '@marigold/system';
 
 export const HelpText: ThemeComponent<'HelpText'> = {
-  container: cva('group-error/field:text-text-error'),
+  container: cva(
+    'group-error/field:text-text-error group-[invalid=true]/field:text-text-error'
+  ),
   icon: cva('h-4 w-4'),
 };